Put progress logic in dedicated class
